+++ sdk/communication/azure-communication-phonenumbers/azure/communication/phonenumbers/_generated/aio/_configuration.py
@@ -6,21 +6,14 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-import sys
 from typing import Any
 
-from azure.core.configuration import Configuration
 from azure.core.pipeline import policies
-
-if sys.version_info >= (3, 8):
-    from typing import Literal  # pylint: disable=no-name-in-module, ungrouped-imports
-else:
-    from typing_extensions import Literal  # type: ignore  # pylint: disable=ungrouped-imports
 
 VERSION = "unknown"
 
 
-class PhoneNumbersClientConfiguration(Configuration):  # pylint: disable=too-many-instance-attributes
+class PhoneNumbersClientConfiguration:  # pylint: disable=too-many-instance-attributes,name-too-long
     """Configuration for PhoneNumbersClient.
 
     Note that all parameters used to create this instance are saved as instance
@@ -29,14 +22,13 @@ class PhoneNumbersClientConfiguration(Configuration):  # pylint: disable=too-man
     :param endpoint: The communication resource, for example
      https://resourcename.communication.azure.com. Required.
     :type endpoint: str
-    :keyword api_version: Api Version. Default value is "2022-12-01". Note that overriding this
-     default value may result in unsupported behavior.
+    :keyword api_version: Api Version. Default value is "2024-03-01-preview". Note that overriding
+     this default value may result in unsupported behavior.
     :paramtype api_version: str
     """
 
     def __init__(self, endpoint: str, **kwargs: Any) -> None:
-        super(PhoneNumbersClientConfiguration, self).__init__(**kwargs)
-        api_version: Literal["2022-12-01"] = kwargs.pop("api_version", "2022-12-01")
+        api_version: str = kwargs.pop("api_version", "2024-03-01-preview")
 
         if endpoint is None:
             raise ValueError("Parameter 'endpoint' must not be None.")
@@ -44,6 +36,7 @@ class PhoneNumbersClientConfiguration(Configuration):  # pylint: disable=too-man
         self.endpoint = endpoint
         self.api_version = api_version
         kwargs.setdefault("sdk_moniker", "phonenumbersclient/{}".format(VERSION))
+        self.polling_interval = kwargs.get("polling_interval", 30)
         self._configure(**kwargs)
 
     def _configure(self, **kwargs: Any) -> None:
@@ -52,7 +45,7 @@ class PhoneNumbersClientConfiguration(Configuration):  # pylint: disable=too-man
         self.proxy_policy = kwargs.get("proxy_policy") or policies.ProxyPolicy(**kwargs)
         self.logging_policy = kwargs.get("logging_policy") or policies.NetworkTraceLoggingPolicy(**kwargs)
         self.http_logging_policy = kwargs.get("http_logging_policy") or policies.HttpLoggingPolicy(**kwargs)
-        self.retry_policy = kwargs.get("retry_policy") or policies.AsyncRetryPolicy(**kwargs)
         self.custom_hook_policy = kwargs.get("custom_hook_policy") or policies.CustomHookPolicy(**kwargs)
         self.redirect_policy = kwargs.get("redirect_policy") or policies.AsyncRedirectPolicy(**kwargs)
+        self.retry_policy = kwargs.get("retry_policy") or policies.AsyncRetryPolicy(**kwargs)
         self.authentication_policy = kwargs.get("authentication_policy")