+++ sdk/communication/azure-communication-phonenumbers/test/test_phone_number_operations.py
@@ -7,10 +7,11 @@
 import pytest
 from azure.core.rest import HttpRequest
 from azure.communication.phonenumbers._generated.operations._operations import (
-build_phone_numbers_get_search_result_request, 
+build_phone_numbers_get_search_result_request,
 build_phone_numbers_purchase_phone_numbers_request,
 build_phone_numbers_get_operation_request,
-build_phone_numbers_cancel_operation_request
+build_phone_numbers_cancel_operation_request,
+build_phone_numbers_operator_information_search_request,
 )
 
 test_id = "test_id"
@@ -21,7 +22,7 @@ def test_build_phone_numbers_get_search_result_request():
     assert isinstance(request, HttpRequest)
     assert request.method == "GET"
     assert test_id in request.url
-    assert "api-version=2022-12-01" in request.url
+    assert "api-version=2024-03-01-preview" in request.url
     assert request.headers["Accept"] == "application/json"
 
 def test_build_phone_numbers_purchase_phone_numbers_request():
@@ -46,4 +47,13 @@ def test_build_phone_numbers_cancel_operation_request():
     assert isinstance(request, HttpRequest)
     assert request.method == "DELETE"
     assert test_id in request.url
+    assert request.headers["Accept"] == "application/json"
+
+def test_build_phone_numbers_operator_information_search_request():
+    request = build_phone_numbers_operator_information_search_request()
+
+    assert isinstance(request, HttpRequest)
+    assert request.method == "POST"
+    assert "api-version=2024-03-01-preview" in request.url
+    assert "operatorInformation/:search" in request.url
     assert request.headers["Accept"] == "application/json"